Reorganize files into component folders
