+++ src/Header/Header.css
@@ -19,6 +19,7 @@ header button{
 }
 
 header h1{
+  height: 100%;
   align-self: center;
   color: #FEFDD0;
   font-size: 4em;
@@ -26,11 +27,11 @@ header h1{
 }
 
 .icon {
-height: 100%
+  height: 100%
 }
 
 .icon-holder {
-display: flex;
-height: 100%;
-width: 100%;
+  display: flex;
+  height: 75%;
+  width: 100%;
 }
